# control-plane/openapi.json
{
  "components": {
    "schemas": {
...
      "DesiredStateResponse": {
        "properties": {
          "deployments": {
            "items": {
              "$ref": "#/components/schemas/DesiredDeployment"
            },
            "title": "Deployments",
            "type": "array"
          },
          "max_generation": {
            "title": "Max Generation",
            "type": "integer"
          },
          "stamp_id": {
            "format": "uuid",
            "title": "Stamp Id",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "stamp_id",
          "max_generation",
          "deployments"
        ],
        "title": "DesiredStateResponse",
        "type": "object"
      },
      "EnrollmentTokenCreateRequest": {
        "properties": {
          "allowed_mode": {
            "default": "byoi",
            "enum": [
              "managed",
              "byoi"
            ],
            "title": "Allowed Mode",
            "type": "string"
          },
          "expires_in_minutes": {
            "default": 60,
            "maximum": 1440.0,
            "minimum": 5.0,
            "title": "Expires In Minutes",
            "type": "integer"
          }
        },
        "title": "EnrollmentTokenCreateRequest",
        "type": "object"
      },
      "EnrollmentTokenResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "allowed_mode": {
            "title": "Allowed Mode",
            "type": "string"
          },
          "enrollment_token": {
            "description": "Shown once and never stored",
            "title": "Enrollment Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "allowed_mode",
          "expires_at",
          "enrollment_token"
        ],
        "title": "EnrollmentTokenResponse",
        "type": "object"
      },
...
      "FabricGpuClaim": {
        "additionalProperties": false,
        "description": "Devices one deployment's model hosts are holding on a stamp.",
        "properties": {
          "deployment_id": {
            "maxLength": 64,
            "title": "Deployment Id",
            "type": "string"
          },
          "gpus": {
            "default": 0,
            "maximum": 1024.0,
            "minimum": 0.0,
            "title": "Gpus",
            "type": "integer"
          }
        },
        "required": [
          "deployment_id"
        ],
        "title": "FabricGpuClaim",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeartbeatRequest": {
        "properties": {
          "capabilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StampCapabilities"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "stamp_id": {
            "format": "uuid",
            "title": "Stamp Id",
            "type": "string"
          }
        },
        "required": [
          "stamp_id",
          "received_at"
        ],
        "title": "HeartbeatResponse",
        "type": "object"
      },
...
      "StampCapabilities": {
        "additionalProperties": false,
        "description": "Bounded capability report.\n\nSecrets, arbitrary labels, node files, and request content are excluded by\nschema. Unknown fields are rejected.",
        "properties": {
          "agent_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "allocatable_gpus": {
            "default": 0,
            "minimum": 0.0,
            "title": "Allocatable Gpus",
            "type": "integer"
          },
          "available_gpu_slots": {
            "items": {
              "type": "integer"
            },
            "maxItems": 8,
            "title": "Available Gpu Slots",
            "type": "array"
          },
          "collector_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Version"
          },
          "container_runtime_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Runtime Version"
          },
          "driver_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Version"
          },
          "fabric_gpu_claims": {
            "items": {
              "$ref": "#/components/schemas/FabricGpuClaim"
            },
            "maxItems": 512,
            "title": "Fabric Gpu Claims",
            "type": "array"
          },
          "fabric_requested_gpus": {
            "default": 0,
            "minimum": 0.0,
            "title": "Fabric Requested Gpus",
            "type": "integer"
          },
          "gpu_claims_measured": {
            "default": false,
            "title": "Gpu Claims Measured",
            "type": "boolean"
          },
          "gpus": {
            "items": {
              "$ref": "#/components/schemas/GpuCapability"
            },
            "maxItems": 64,
            "title": "Gpus",
            "type": "array"
          },
          "inference_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inference Url"
          },
          "max_free_gpus_per_node": {
            "default": 0,
            "minimum": 0.0,
            "title": "Max Free Gpus Per Node",
            "type": "integer"
          },
          "max_gpus_per_node": {
            "default": 0,
            "minimum": 0.0,
            "title": "Max Gpus Per Node",
            "type": "integer"
          },
          "operator_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Version"
          },
          "orchestrator": {
            "maxLength": 32,
            "title": "Orchestrator",
            "type": "string"
          },
          "orchestrator_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orchestrator Version"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "requested_gpus": {
            "default": 0,
            "minimum": 0.0,
            "title": "Requested Gpus",
            "type": "integer"
          },
          "runtime_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Version"
          }
        },
        "required": [
          "orchestrator"
        ],
        "title": "StampCapabilities",
        "type": "object"
      },
      "StampEnrollRequest": {
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/StampCapabilities"
          },
          "enrollment_token": {
            "minLength": 8,
            "title": "Enrollment Token",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "enrollment_token",
          "name",
          "capabilities"
        ],
        "title": "StampEnrollRequest",
        "type": "object"
      },
      "StampEnrollResponse": {
        "properties": {
          "agent_credential": {
            "description": "Shown once; agent-only Secret",
            "title": "Agent Credential",
            "type": "string"
          },
          "stamp": {
            "$ref": "#/components/schemas/StampResponse"
          },
          "telemetry_credential": {
            "description": "Shown once; collector-only Secret",
            "title": "Telemetry Credential",
            "type": "string"
          }
        },
        "required": [
          "stamp",
          "agent_credential",
          "telemetry_credential"
        ],
        "title": "StampEnrollResponse",
        "type": "object"
      },
      "StampResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "capabilities": {
            "title": "Capabilities",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "orchestrator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orchestrator"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "name",
          "mode",
          "orchestrator",
          "region",
          "status",
          "capabilities",
          "last_heartbeat_at",
          "revoked_at",
          "created_at"
        ],
        "title": "StampResponse",
        "type": "object"
      },
...
      "StatusReportRequest": {
        "properties": {
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/StatusCondition"
            },
            "maxItems": 32,
            "title": "Conditions",
            "type": "array"
          },
          "deployment_id": {
            "format": "uuid",
            "title": "Deployment Id",
            "type": "string"
          },
          "endpoint": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "observed_generation": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Generation"
          },
          "phase": {
            "enum": [
              "pending",
              "progressing",
              "ready",
              "degraded",
              "failed",
              "terminating"
            ],
            "title": "Phase",
            "type": "string"
          },
          "ready_replicas": {
            "default": 0,
            "minimum": 0.0,
            "title": "Ready Replicas",
            "type": "integer"
          },
          "unavailable_replicas": {
            "default": 0,
            "minimum": 0.0,
            "title": "Unavailable Replicas",
            "type": "integer"
          }
        },
        "required": [
          "deployment_id",
          "phase"
        ],
        "title": "StatusReportRequest",
        "type": "object"
      },
      "StatusReportResponse": {
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "deployment_id": {
            "format": "uuid",
            "title": "Deployment Id",
            "type": "string"
          },
          "stamp_id": {
            "format": "uuid",
            "title": "Stamp Id",
            "type": "string"
          }
        },
        "required": [
          "deployment_id",
          "stamp_id",
          "accepted_at"
        ],
        "title": "StatusReportResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Accounts, credentials, deployment intent, and inference stamps.",
    "title": "Fabric Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/accounts/{account_id}/stamp-enrollment-tokens": {
      "post": {
        "operationId": "create_token_v1_accounts__account_id__stamp_enrollment_tokens_post",
        "parameters": [
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentTokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a single-use stamp enrollment token",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/accounts/{account_id}/stamp-enrollment-tokens/{token_id}": {
      "delete": {
        "operationId": "revoke_token_v1_accounts__account_id__stamp_enrollment_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Token Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke an unused stamp enrollment token",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/accounts/{account_id}/stamps": {
      "get": {
        "operationId": "list_account_stamps_v1_accounts__account_id__stamps_get",
        "parameters": [
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StampResponse"
                  },
                  "title": "Response List Account Stamps V1 Accounts  Account Id  Stamps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List inference stamps owned by this account",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/accounts/{account_id}/stamps/{stamp_id}": {
      "delete": {
        "description": "Stop further synchronization for this stamp.\n\nThe agent and telemetry credentials are revoked in the same transaction, so\nheartbeat, desired-state, status, and future telemetry export are all denied.",
        "operationId": "revoke_account_stamp_v1_accounts__account_id__stamps__stamp_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "stamp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stamp Id",
              "type": "string"
            }
          },
          {
            "description": "Account identifier",
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "description": "Account identifier",
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke a stamp and its machine credentials",
        "tags": [
          "stamps"
        ]
      }
    },
...
    "/v1/stamps/enroll": {
      "post": {
        "description": "Account ownership is derived from the enrollment record, never the payload.",
        "operationId": "enroll_v1_stamps_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StampEnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampEnrollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a stamp using a one-time enrollment token",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/stamps/{stamp_id}/desired-state": {
      "get": {
        "operationId": "desired_state_v1_stamps__stamp_id__desired_state_get",
        "parameters": [
          {
            "in": "path",
            "name": "stamp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stamp Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_generation",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Generation",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DesiredStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pull assignments newer than the acknowledged generation",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/stamps/{stamp_id}/heartbeat": {
      "post": {
        "operationId": "heartbeat_v1_stamps__stamp_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "stamp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stamp Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report liveness and bounded capabilities",
        "tags": [
          "stamps"
        ]
      }
    },
    "/v1/stamps/{stamp_id}/status": {
      "post": {
        "operationId": "report_v1_stamps__stamp_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "stamp_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Stamp Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report observed deployment status",
        "tags": [
          "stamps"
        ]
      }
    },
...
  }
}